rename UserInterfact to MessageSender and moved Chatbot interface to under chatbot directory
