+++ src/main/java/codeu/controller/ChatServlet.java
@@ -16,10 +16,10 @@ package codeu.controller;
 
 import codeu.model.data.Conversation;
 import codeu.model.data.Message;
-import codeu.model.data.user.Chatbot;
+import codeu.model.data.user.chatbot.Chatbot;
 import codeu.model.data.user.User;
 import codeu.model.data.user.UserGroup;
-import codeu.model.data.user.UserInterface;
+import codeu.model.data.user.MessageSender;
 import codeu.model.data.user.chatbot.HelloChatbot;
 import codeu.model.store.basic.ConversationStore;
 import codeu.model.store.basic.MessageStore;
@@ -164,7 +164,7 @@ public class ChatServlet extends HttpServlet {
   }
 
   public void sendMessageToConversation(
-      UserInterface user,
+      MessageSender user,
       String messageContent,
       Conversation conversation)
   {
