--- src/main/java/codeu/model/data/user/Chatbot.java
@@ -1,7 +1,0 @@
-package codeu.model.data.user;
-
-public interface Chatbot extends UserInterface {
-
-  String respondToMessageFrom(User sender, String messageContent);
-
-}
