+++ src/main/java/codeu/model/data/user/chatbot/Chatbot.java
@@ -1,0 +1,10 @@
+package codeu.model.data.user.chatbot;
+
+import codeu.model.data.user.MessageSender;
+import codeu.model.data.user.User;
+
+public interface Chatbot extends MessageSender {
+
+  String respondToMessageFrom(User sender, String messageContent);
+
+}
