+++ src/test/java/codeu/controller/ChatServletTest.java
@@ -16,7 +16,7 @@ package codeu.controller;
 
 import codeu.model.data.Conversation;
 import codeu.model.data.Message;
-import codeu.model.data.user.Chatbot;
+import codeu.model.data.user.chatbot.Chatbot;
 import codeu.model.data.user.User;
 import codeu.model.data.user.UserGroup;
 import codeu.model.store.basic.ConversationStore;
